api/requirements.txt
flask
python-dotenv
psycopg2-binary
sqlalchemy
flask-sqlalchemy
flask-cors
python-jose
passlib
requests
google-genai
flasgger
marshmallow
apispec